Add plots of data
Looking at the data like this is quite important. It helped me figure out
which of the sensors is which. It also provides some ideas on how the
algorithm could be implemented

Also included standard deviation. I'm thinking about including standard
deviation in some way on mysolution, not entirely sure how at the moment