Implement saving of past data in a numpy 3D array
On a cell at the bottom of solution.py, I implemented some funcions
that record data of the past N iterations in an array. I also
implemented a function to calculate the mean for each sensor and value
with this data.
I will now integrate what I have in this cell with the rest of the code